+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 # Changelog
 
+### - 2024-10-18
+- env/env.php composer.json autoload 등록 및 테스트
+
 ### - 2024-10-17
 - 내 네임스페이스와 경로 composer.json 설정 및 라우터 실행 관련 업데이트 
 - My\Service\Test\Test class example update